requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
chromadb==0.4.18
sentence-transformers==2.2.2
PyPDF2==3.0.1
Pillow==10.1.0
pytesseract==0.3.10
python-dotenv==1.0.0
groq==0.4.2
aiofiles==23.2.1
